Rename UserOverviewStore file name
