+++ apps/authentication-server-ui/src/app/user/overview/actions/edit-user-button/edit-user-button.component.spec.ts
@@ -7,7 +7,7 @@ import {
 } from '@dnd-mapp/authentication-server-ui/testing';
 import { defaultUsers } from '@dnd-mapp/data';
 import { provideTranslations } from '../../../../shared';
-import { UsersOverviewStore } from '../../../services/users-overview-store';
+import { UsersOverviewStore } from '../../../services/users-overview.store';
 import { EditUserButtonComponent } from './edit-user-button.component';
 
 describe('EditUserButtonComponent', () => {
